files included to the repo/modified during CPU color deconvolution optimization

+++ src/features-cpu/Operators.cpp
@@ -10,12 +10,12 @@
 // Assuming n > 0
 int Operators::rndint(float n)//round float to the nearest integer
 {	
-	int ret = floor(n);
+	int ret = (int)floor(n);
 	float t;
 	t=n-floor(n);
 	if (t>=0.5)    
 	{
-		ret = floor(n) + 1;
+		ret = (int)floor(n) + 1;
 	}
 	return ret;
 }
